+++ peek_core_search/_private/storage/SearchObjectCompilerQueue.py
@@ -1,12 +1,12 @@
 import logging
 
-from sqlalchemy import Column
-from sqlalchemy import Integer
-from vortex.Tuple import Tuple, addTupleType
-
 from peek_abstract_chunked_index.private.tuples.ACIProcessorQueueTupleABC import \
     ACIProcessorQueueTupleABC
 from peek_core_search._private.PluginNames import searchTuplePrefix
+from sqlalchemy import Column, BigInteger
+from sqlalchemy import Integer
+from vortex.Tuple import Tuple, addTupleType
+
 from .DeclarativeBase import DeclarativeBase
 
 logger = logging.getLogger(__name__)
@@ -18,7 +18,7 @@ class SearchObjectCompilerQueue(Tuple, DeclarativeBase,
     __tablename__ = 'SearchObjectCompilerQueue'
     __tupleType__ = searchTuplePrefix + 'SearchObjectCompilerQueueTable'
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     chunkKey = Column(Integer, primary_key=True)
 
     @classmethod
@@ -28,5 +28,3 @@ class SearchObjectCompilerQueue(Tuple, DeclarativeBase,
     @property
     def ckiUniqueKey(self):
         return self.chunkKey
-
-
